+++ src/Java/myservicebus/src/main/java/com/myservicebus/topology/BusTopology.java
@@ -5,5 +5,8 @@ import java.util.List;
 public interface BusTopology {
     List<MessageTopology> getMessages();
     List<ConsumerTopology> getConsumers();
-}
 
+    default TopologySnapshot getSnapshot() {
+        return TopologySnapshots.create(this);
+    }
+}