+++ src/MyServiceBus/Topology/IBusTopology.cs
@@ -6,5 +6,6 @@ public interface IBusTopology
 {
     List<MessageTopology> Messages { get; }
     List<ConsumerTopology> Consumers { get; }
-}
 
+    TopologySnapshot GetSnapshot() => TopologySnapshotBuilder.Create(this);
+}